+++ CHANGELOG.md
@@ -11,8 +11,8 @@ Keep it human-readable, your future self will thank you!
 ## [Unreleased]
 
 ### Added
-- Fix: Enable inference when no constant forcings are used
 
+- Fix: Enable inference when no constant forcings are used
 - Add anemoi-transform link to documentation
 
 ### Changed
@@ -20,6 +20,7 @@ Keep it human-readable, your future self will thank you!
 - Add cos_solar_zenith_angle to list of known forcings
 - Add missing classes in checkpoint handling
 - Rename Condition to State [#24](https://github.com/ecmwf/anemoi-inference/pull/24)
+- Fix pre-commit regex
 
 ### Removed